StartNoLightOnlyDebt/back/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
bs4==0.0.2
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
dj-rest-auth==7.0.0
Django==4.2.16
django-allauth==65.2.0
django-cors-headers==4.6.0
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-seed==0.3.1
djangorestframework==3.15.1
exceptiongroup==1.2.2
executing==2.1.0
Faker==30.4.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
fonttools==4.53.1
gevent==24.10.3
geventhttpclient==2.3.1
greenlet==3.1.1
h11==0.14.0
idna==3.7
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
kiwisolver==1.4.7
locust==2.32.1
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
msgpack==1.1.0
numpy==2.0.2
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
prompt_toolkit==3.0.47
psutil==6.1.0
psycopg2==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2023.3.post1
pywin32==308
pyzmq==26.2.0
requests==2.31.0
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.0
stack-data==0.6.3
tomli==2.0.2
toposort==1.10
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.2
wsproto==1.2.0
zipp==3.20.2
zope.event==5.0
zope.interface==7.1.1
